Refactor

+++ Core/Money.cs
@@ -2,13 +2,13 @@ namespace Core
 {
 	public struct Money
 	{
-		public Money(double amount, Currency currency)
+		public Money(decimal amount, Currency currency)
 		{
 			Amount = amount;
 			Currency = currency;
 		}
 
-		public double Amount { get; set; }
+		public decimal Amount { get; set; }
 		public Currency Currency { get; set; }
 	}
 }
